pyproject.toml
[project]
name = "drama-catch-up"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["requests", "pyyaml", "beautifulsoup4", "browser-cookie3"]

[project.optional-dependencies]
build = ["pyinstaller", "ruff"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "C4", "UP", "SIM", "ASYNC"]

[tool.ruff.format]
quote-style = "double"

[dependency-groups]
dev = ["pyinstaller>=6.15.0", "ruff>=0.12.11"]
